requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
pandas==2.1.3
yfinance==0.2.28
requests-cache==1.1.1
requests==2.31.0
python-dateutil==2.8.2
rapidfuzz==3.5.2
reportlab==4.0.7
python-multipart==0.0.6
python-dotenv==1.0.0
mammoth==1.6.0
PyMuPDF==1.23.8
Pillow==10.1.0
plotly==5.18.0

